+++ Plugins/Events/Events/SpellEvents.cpp
@@ -4,11 +4,33 @@
 #include "API/CNWSEffectListHandler.hpp"
 #include "API/CNWRules.hpp"
 #include "API/CNWSpellArray.hpp"
+#include "API/CNWCCMessageData.hpp"
 
 namespace Events {
 
 using namespace NWNXLib;
 using namespace NWNXLib::API;
+
+struct SpellCastAction
+{
+    uint32_t nSpellId;
+    uint8_t nMultiClass;
+    uint32_t nDomainLevel;
+    uint32_t nMetaMagic;
+    BOOL bSpontaneous;
+    BOOL bDefensiveCast;
+    OBJECT_ID oidTarget;
+    float fTargetX;
+    float fTargetY;
+    float fTargetZ;
+    BOOL bInstant;
+    uint32_t nProjectilePathType;
+    uint32_t nFeat;
+    uint8_t nCasterLevel;
+    BOOL bFake;
+};
+
+static SpellCastAction LastSpellAction;
 
 static Hooks::Hook s_SpellCastAndImpactHook;
 static Hooks::Hook s_SetMemorizedSpellSlotHook;
@@ -16,6 +38,10 @@ static Hooks::Hook s_ClearMemorizedSpellSlotHook;
 static Hooks::Hook s_BroadcastSpellCastHook;
 static Hooks::Hook s_OnEffectAppliedHook;
 static Hooks::Hook s_DecrementSpellReadyCountHook;
+static Hooks::Hook s_ClearActionHook;
+static Hooks::Hook s_AIActionCastSpellHook;
+static Hooks::Hook s_SendFeedbackMessageHook;
+static Hooks::Hook s_BroadcastCounterSpellDataHook;
 
 static void SpellCastAndImpactHook(CNWSObject*, uint32_t, Vector, ObjectID, uint8_t, ObjectID, int32_t, int32_t, uint8_t, int32_t);
 static int32_t SetMemorizedSpellSlotHook(CNWSCreatureStats*, uint8_t, uint8_t, uint32_t, uint8_t, uint8_t, int32_t);
@@ -23,6 +49,11 @@ static void ClearMemorizedSpellSlotHook(CNWSCreatureStats*, uint8_t, uint8_t, ui
 static void BroadcastSpellCastHook(CNWSCreature*, uint32_t, uint8_t, uint16_t);
 static int32_t OnEffectAppliedHook(CNWSEffectListHandler*, CNWSObject*, CGameEffect*, int32_t);
 static int32_t DecrementSpellReadyCountHook(CNWSCreature*, uint32_t, uint8_t, uint8_t, uint8_t, uint8_t);
+static BOOL ClearActionHook(CNWSCreature*, CNWSObjectActionNode*, BOOL);
+static BOOL ClearActionHook(CNWSCreature*, CNWSObjectActionNode*, BOOL);
+static uint32_t AIActionCastSpellHook(CNWSCreature*, CNWSObjectActionNode*);
+static void SendFeedbackMessageHook(CNWSCreature*, uint16_t, CNWCCMessageData*, CNWSPlayer*);
+static void BroadcastCounterSpellDataHook(CNWSObject*, CNWSpell*, CNWCCMessageData*);
 
 void SpellEvents() __attribute__((constructor));
 void SpellEvents()
@@ -55,6 +86,13 @@ void SpellEvents()
     InitOnFirstSubscribe("NWNX_ON_DECREMENT_SPELL_COUNT_.*", []() {
         s_DecrementSpellReadyCountHook = Hooks::HookFunction(&CNWSCreature::DecrementSpellReadyCount,
                                              &DecrementSpellReadyCountHook, Hooks::Order::Early);
+    });
+
+    InitOnFirstSubscribe("NWNX_ON_SPELL_FAILED_.*", []() {
+        s_AIActionCastSpellHook = Hooks::HookFunction(&CNWSCreature::AIActionCastSpell, &AIActionCastSpellHook, Hooks::Order::Earliest);
+        s_ClearActionHook = Hooks::HookFunction(Functions::_ZN12CNWSCreature11ClearActionEP20CNWSObjectActionNodei, &ClearActionHook, Hooks::Order::Early);
+        s_SendFeedbackMessageHook = Hooks::HookFunction(&CNWSCreature::SendFeedbackMessage, &SendFeedbackMessageHook, Hooks::Order::Early);
+        s_BroadcastCounterSpellDataHook = Hooks::HookFunction(&CNWSObject::BroadcastCounterSpellData, &BroadcastCounterSpellDataHook, Hooks::Order::Early);
     });
 }
 
@@ -249,5 +287,155 @@ int32_t DecrementSpellReadyCountHook(CNWSCreature *thisPtr, uint32_t nSpellID, u
     return retVal;
 }
 
+uint32_t AIActionCastSpellHook(CNWSCreature* thisPtr, CNWSObjectActionNode* pNode)
+{
+    LastSpellAction.nSpellId = pNode->m_pParameter[0];
+    LastSpellAction.nMultiClass = pNode->m_pParameter[1] & 0xFF;
+    LastSpellAction.nDomainLevel = pNode->m_pParameter[2];
+    LastSpellAction.nMetaMagic = pNode->m_pParameter[3];
+    LastSpellAction.bSpontaneous = pNode->m_pParameter[4] != 0;
+    LastSpellAction.bDefensiveCast = (pNode->m_pParameter[1] & 0x0000FF00) >> 8;
+    LastSpellAction.oidTarget = pNode->m_pParameter[5];
+    LastSpellAction.fTargetX = *((float*)&pNode->m_pParameter[6]);
+    LastSpellAction.fTargetY = *((float*)&pNode->m_pParameter[7]);
+    LastSpellAction.fTargetZ = *((float*)&pNode->m_pParameter[8]);
+    LastSpellAction.bInstant = (pNode->m_pParameter[9] & 0x40000000) != 0;
+    LastSpellAction.nProjectilePathType = pNode->m_pParameter[9] & 0x3FFFFFFF;
+    LastSpellAction.nFeat = pNode->m_pParameter[10];
+    LastSpellAction.nCasterLevel = pNode->m_pParameter[11] & 0x000000FF;
+    LastSpellAction.bFake = (pNode->m_pParameter[9] & 0x80000000) != 0;
+
+    return s_AIActionCastSpellHook->CallOriginal<uint32_t>(thisPtr, pNode);
+}
+
+BOOL ClearActionHook(CNWSCreature* thisPtr, CNWSObjectActionNode* pNode, BOOL bIsTopmostAction)
+{
+    if ((IsIDInWhitelist("NWNX_ON_SPELL_FAILED", pNode->m_pParameter[0])) && 
+        (bIsTopmostAction) && (pNode->m_nActionId == 15) && (pNode->m_nParameters == 12) && (pNode->m_bInterruptable))
+    {
+        auto PushAndSignal = [&](const std::string& ev) -> bool {
+            PushEventData("SPELL_ID", std::to_string(pNode->m_pParameter[0]));
+            PushEventData("MULTI_CLASS", std::to_string(pNode->m_pParameter[1] & 0xFF));
+            PushEventData("DOMAIN", std::to_string(pNode->m_pParameter[2]));
+            PushEventData("METAMAGIC", std::to_string(pNode->m_pParameter[3]));
+            PushEventData("SPELL_SPONTANEOUS", std::to_string(pNode->m_pParameter[4]));
+            PushEventData("DEFENSIVELY_CAST", std::to_string((pNode->m_pParameter[1] & 0x0000FF00) >> 8));
+            PushEventData("TARGET_OBJECT_ID", Utils::ObjectIDToString(pNode->m_pParameter[5]));
+            PushEventData("TARGET_POSITION_X", std::to_string(*((float*)&pNode->m_pParameter[6])));
+            PushEventData("TARGET_POSITION_Y", std::to_string(*((float*)&pNode->m_pParameter[7])));
+            PushEventData("TARGET_POSITION_Z", std::to_string(*((float*)&pNode->m_pParameter[8])));
+            PushEventData("IS_INSTANT_SPELL", std::to_string((uint32_t)((pNode->m_pParameter[9] & 0x40000000) != 0)));
+            PushEventData("PROJECTILE_PATH_TYPE", std::to_string(pNode->m_pParameter[9] & 0x3FFFFFFF));
+            PushEventData("FEAT", std::to_string(pNode->m_pParameter[10]));
+            PushEventData("CASTERLEVEL", std::to_string(pNode->m_pParameter[11] & 0x000000FF));
+            PushEventData("IS_FAKE", std::to_string((uint32_t)((pNode->m_pParameter[9] & 0x80000000) != 0)));
+            PushEventData("REASON", "0" /* NWNX_EVENTS_SPELLFAIL_REASON_CANCELED */);
+            return SignalEvent(ev, thisPtr->m_idSelf);
+        };
+
+        PushAndSignal("NWNX_ON_SPELL_FAILED_BEFORE");
+
+        auto retVal = s_ClearActionHook->CallOriginal<BOOL>(thisPtr, pNode, bIsTopmostAction);
+
+        PushAndSignal("NWNX_ON_SPELL_FAILED_AFTER");
+
+        return retVal;
+    }
+    
+    return s_ClearActionHook->CallOriginal<BOOL>(thisPtr, pNode, bIsTopmostAction);
+}
+
+void BroadcastCounterSpellDataHook(CNWSObject* thisPtr, CNWSpell* pSpell, CNWCCMessageData* pMessageData)
+{
+    if (!IsIDInWhitelist("NWNX_ON_SPELL_FAILED", LastSpellAction.nSpellId))
+    {
+        s_BroadcastCounterSpellDataHook->CallOriginal<void>(thisPtr, pSpell, pMessageData);
+        return;
+    }
+
+    auto PushAndSignal = [&](const std::string& ev) -> bool {
+        PushEventData("SPELL_ID", std::to_string(LastSpellAction.nSpellId));
+        PushEventData("MULTI_CLASS", std::to_string(LastSpellAction.nMultiClass));
+        PushEventData("DOMAIN", std::to_string(LastSpellAction.nDomainLevel));
+        PushEventData("METAMAGIC", std::to_string(LastSpellAction.nMetaMagic));
+        PushEventData("SPELL_SPONTANEOUS", std::to_string((uint32_t)LastSpellAction.bSpontaneous));
+        PushEventData("DEFENSIVELY_CAST", std::to_string((uint32_t)LastSpellAction.bDefensiveCast));
+        PushEventData("TARGET_OBJECT_ID", Utils::ObjectIDToString(LastSpellAction.oidTarget));
+        PushEventData("TARGET_POSITION_X", std::to_string(LastSpellAction.fTargetX));
+        PushEventData("TARGET_POSITION_Y", std::to_string(LastSpellAction.fTargetY));
+        PushEventData("TARGET_POSITION_Z", std::to_string(LastSpellAction.fTargetZ));
+        PushEventData("IS_INSTANT_SPELL", std::to_string((uint32_t)LastSpellAction.bInstant));
+        PushEventData("PROJECTILE_PATH_TYPE", std::to_string(LastSpellAction.nProjectilePathType));
+        PushEventData("FEAT", std::to_string(LastSpellAction.nFeat));
+        PushEventData("CASTERLEVEL", std::to_string(LastSpellAction.nCasterLevel));
+        PushEventData("IS_FAKE", std::to_string((uint32_t)LastSpellAction.bFake));
+        PushEventData("REASON", "1" /* NWNX_EVENTS_SPELLFAIL_REASON_COUNTERSPELL */);
+        return SignalEvent(ev, thisPtr->m_idSelf);
+    };
+
+    PushAndSignal("NWNX_ON_SPELL_FAILED_BEFORE");
+
+    s_BroadcastCounterSpellDataHook->CallOriginal<void>(thisPtr, pSpell, pMessageData);
+
+    PushAndSignal("NWNX_ON_SPELL_FAILED_AFTER");  
+}
+
+void SendFeedbackMessageHook(CNWSCreature* thisPtr, uint16_t nFeedbackID, CNWCCMessageData* pMessageData, CNWSPlayer* pFeedbackPlayer)
+{
+    if ((IsIDInWhitelist("NWNX_ON_SPELL_FAILED", LastSpellAction.nSpellId)) &&
+        (
+            (nFeedbackID == 61 /* NWNX_FEEDBACK_CAST_ARCANE_SPELL_FAILURE */) || 
+            (nFeedbackID == 236 /* NWNX_FEEDBACK_CAST_EFFECT_SPELL_FAILURE */) || 
+            (nFeedbackID == 21 /* NWNX_FEEDBACK_CAST_LOST_TARGET */) || 
+            (nFeedbackID == 211 /* NWNX_FEEDBACK_CAST_USE_MOUTH */) || 
+            (nFeedbackID == 233 /* NWNX_FEEDBACK_CAST_DEFCAST_CONCENTRATION_FAILURE */) ||
+            (nFeedbackID == 65 /* NWNX_FEEDBACK_CAST_ENTANGLE_CONCENTRATION_FAILURE */) ||
+            // Untested:
+            (nFeedbackID == 107 /* NWNX_FEEDBACK_CAST_CANT_CAST_WHILE_POLYMORPHED */) ||
+            (nFeedbackID == 22 /* NWNX_FEEDBACK_CAST_CANT_CAST */) || 
+            (nFeedbackID == 210 /* NWNX_FEEDBACK_CAST_USE_HANDS */)
+        ))
+    {
+        std::unordered_map<uint16_t, uint32_t> feedback2Reason = {
+            {61 /* NWNX_FEEDBACK_CAST_ARCANE_SPELL_FAILURE */,           2  /* NWNX_EVENTS_SPELLFAIL_REASON_ASF */},
+            {236 /* NWNX_FEEDBACK_CAST_EFFECT_SPELL_FAILURE */,          3  /* NWNX_EVENTS_SPELLFAIL_REASON_SPELLFAILURE */}, 
+            {21 /* NWNX_FEEDBACK_CAST_LOST_TARGET */,                    4  /* NWNX_EVENTS_SPELLFAIL_REASON_LOST_TARGET */},
+            {211 /* NWNX_FEEDBACK_CAST_USE_MOUTH */,                     5  /* NWNX_EVENTS_SPELLFAIL_REASON_SILENCED */}, 
+            {233 /* NWNX_FEEDBACK_CAST_DEFCAST_CONCENTRATION_FAILURE */, 6  /* NWNX_EVENTS_SPELLFAIL_REASON_DEFCAST_CONCENTRATION */}, 
+            {65 /* NWNX_FEEDBACK_CAST_ENTANGLE_CONCENTRATION_FAILURE */, 7  /* NWNX_EVENTS_SPELLFAIL_REASON_ENTANGLE_CONCENTRATION */}, 
+            {107 /* NWNX_FEEDBACK_CAST_CANT_CAST_WHILE_POLYMORPHED */,   8  /* NWNX_EVENTS_SPELLFAIL_REASON_POLYMORPHED */}, 
+            {22 /* NWNX_FEEDBACK_CAST_CANT_CAST */,                      9  /* NWNX_EVENTS_SPELLFAIL_REASON_CANT_CAST */}, 
+            {210 /* NWNX_FEEDBACK_CAST_USE_HANDS */,                     10 /* NWNX_EVENTS_SPELLFAIL_REASON_CANT_USE_HANDS*/}, 
+        };  
+
+        auto PushAndSignal = [&](const std::string& ev) -> bool {
+            PushEventData("SPELL_ID", std::to_string(LastSpellAction.nSpellId));
+            PushEventData("MULTI_CLASS", std::to_string(LastSpellAction.nMultiClass));
+            PushEventData("DOMAIN", std::to_string(LastSpellAction.nDomainLevel));
+            PushEventData("METAMAGIC", std::to_string(LastSpellAction.nMetaMagic));
+            PushEventData("SPELL_SPONTANEOUS", std::to_string((uint32_t)LastSpellAction.bSpontaneous));
+            PushEventData("DEFENSIVELY_CAST", std::to_string((uint32_t)LastSpellAction.bDefensiveCast));
+            PushEventData("TARGET_OBJECT_ID", Utils::ObjectIDToString(LastSpellAction.oidTarget));
+            PushEventData("TARGET_POSITION_X", std::to_string(LastSpellAction.fTargetX));
+            PushEventData("TARGET_POSITION_Y", std::to_string(LastSpellAction.fTargetY));
+            PushEventData("TARGET_POSITION_Z", std::to_string(LastSpellAction.fTargetZ));
+            PushEventData("IS_INSTANT_SPELL", std::to_string((uint32_t)LastSpellAction.bInstant));
+            PushEventData("PROJECTILE_PATH_TYPE", std::to_string(LastSpellAction.nProjectilePathType));
+            PushEventData("FEAT", std::to_string(LastSpellAction.nFeat));
+            PushEventData("CASTERLEVEL", std::to_string(LastSpellAction.nCasterLevel));
+            PushEventData("IS_FAKE", std::to_string((uint32_t)LastSpellAction.bFake));
+            PushEventData("REASON", std::to_string(feedback2Reason[nFeedbackID]));
+            return SignalEvent(ev, thisPtr->m_idSelf);
+        };
+
+        PushAndSignal("NWNX_ON_SPELL_FAILED_BEFORE");
+
+        s_SendFeedbackMessageHook->CallOriginal<void>(thisPtr, nFeedbackID, pMessageData, pFeedbackPlayer);
+
+        PushAndSignal("NWNX_ON_SPELL_FAILED_AFTER");
+    }
+    else
+        s_SendFeedbackMessageHook->CallOriginal<void>(thisPtr, nFeedbackID, pMessageData, pFeedbackPlayer);
+}
 
 }